refactor: Set up layout and styling for team page (#4172)
