+++ editor.planx.uk/src/pages/layout/AuthenticatedLayout.tsx
@@ -24,7 +24,7 @@ const DashboardContainer = styled(Box)(({ theme }) => ({
   flexDirection: "row",
   width: "100%",
   overflow: "hidden",
-  [`& > .${containerClasses.root}`]: {
+  [`& > .${containerClasses.root}, & > div > .${containerClasses.root}`]: {
     paddingTop: theme.spacing(3),
     paddingBottom: theme.spacing(3),
     [theme.breakpoints.up("lg")]: {
